requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
crispy-bootstrap5==0.6
dj-database-url==1.0.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-jazzmin==2.5.0
django-templated-mail==1.1.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
python-decouple==3.6
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
